Start to integrate controllable steps

+++ bin/handler.sh
@@ -52,22 +52,54 @@ set "$s3_url"
 set "$chamber"
 set "$type"
 set "$committee"
+set "$step_all"
+set "$step_download"
+set "$step_screenshots"
+set "$step_crop_chyrons"
+set "$step_ocr_chyrons"
+set "$step_create_clips"
+set "$step_get_captions"
+set "$step_save_captions"
+set "$step_internet_archive"
+
+if [ "$step_all" = true ]; then
+	step_download=true
+	step_screenshots=true
+	step_crop_chyrons=true
+	step_ocr_chyrons=true
+	step_create_clips=true
+	step_get_captions=true
+	step_save_captions=true
+	step_internet_archive=true
+fi
+
+# If no steps are defined, this is a legacy SQS message -- complete all steps.
+if [ "$step_download" = false ] && [ "$step_screenshots" = false ] && [ "$step_crop_chyrons" = false ] \
+	 && [ "$step_ocr_chyrons" = false ] && [ "$step_create_clips" = false ] \
+	 && [ "$step_get_captions" = false ] && [ "$step_save_captions" = false ] \
+	 && [ "$step_internet_archive" = false ]
+then
+	step_all=true
+fi
 
 # Define the name of the directory that will store the extracted chyrons.
 export output_dir="${filename/.mp4/}"
 
 # OCR the video. This also generates screeshots and thumbnails.
-../bin/ocr.sh "$filename" "$chamber" "$committee" || exit $?
+if [ "$step_screenshots" = true ]; then
+	../bin/ocr.sh "$filename" "$chamber" "$committee" || exit $?
 
-# Move screenshots to S3.
-cd "$VIDEO_DIR" || exit $?
-cd "$output_dir" || exit $?
-if aws s3 sync . s3://video.richmondsunlight.com/"$chamber"/floor/"$date" --exclude "*" --include "*.jpg"
-then
-	echo Deleting all local screenshots
-	rm ./*.jpg
-else
-	echo "AWS S3 sync didn't finish successfully, so screenshots were not thumbnailed or deleted"
+	# Move screenshots to S3.
+	cd "$VIDEO_DIR" || exit $?
+	cd "$output_dir" || exit $?
+	if aws s3 sync . s3://video.richmondsunlight.com/"$chamber"/floor/"$date" --exclude "*" --include "*.jpg"
+	then
+		echo Deleting all local screenshots
+		rm ./*.jpg
+	else
+		echo "AWS S3 sync didn't finish successfully, so screenshots were not thumbnailed or deleted"
+	fi
+
 fi
 
 # Create the record for this video in the database.
@@ -83,7 +115,7 @@ else
 fi
 
 # Only deal with chyrons and captions for floor video.
-if [ "$type" = "floor" ]; then
+if [ "$type" = "floor" ] && [ "$step_ocr_chyrons" = true ]; then
 
 	# Insert the chyrons into the database.
 	php ../bin/save_chyrons.php "$VIDEO_ID" "$output_dir" || exit $?
